--- rules-tests/Restoration/Rector/ClassLike/UpdateFileNameByClassNameFileSystemRector/FixtureUsingMultipleFilesChangedTrait/skip_different_class_name.php.inc
@@ -1,25 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Rector\Tests\Restoration\Rector\ClassLike\UpdateFileNameByClassNameFileSystemRector\FixtureUsingMultipleFilesChangedTrait;
-
-final class SkipDifferentClassName
-{
-
-}
------
------
-SkipDifferentClassName.php
------
------
-<?php
-
-declare(strict_types=1);
-
-namespace Rector\Tests\Restoration\Rector\ClassLike\UpdateFileNameByClassNameFileSystemRector\FixtureUsingMultipleFilesChangedTrait;
-
-final class SkipDifferentClassName
-{
-
-}
