+++ rules-tests/Restoration/Rector/ClassLike/UpdateFileNameByClassNameFileSystemRectorUsingMultipleFilesChangedTrait/Fixture/skip_different_class_name.php.inc
@@ -1,0 +1,25 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rector\Tests\Restoration\Rector\ClassLike\UpdateFileNameByClassNameFileSystemRectorUsingMultipleFilesChangedTrait\Fixture;
+
+final class SkipDifferentClassName
+{
+
+}
+-----
+-----
+SkipDifferentClassName.php
+-----
+-----
+<?php
+
+declare(strict_types=1);
+
+namespace Rector\Tests\Restoration\Rector\ClassLike\UpdateFileNameByClassNameFileSystemRectorUsingMultipleFilesChangedTrait\Fixture;
+
+final class SkipDifferentClassName
+{
+
+}
