+++ rules-tests/Restoration/Rector/ClassLike/UpdateFileNameByClassNameFileSystemRectorUsingMultipleFilesChangedTrait/UpdateFileNameByClassNameFileSystemRectorUsingMultipleFilesChangedTraitTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Rector\Tests\Restoration\Rector\ClassLike\UpdateFileNameByClassNameFileSystemRector;
+namespace Rector\Tests\Restoration\Rector\ClassLike\UpdateFileNameByClassNameFileSystemRectorUsingMultipleFilesChangedTrait;
 
 use Iterator;
 use Rector\Testing\PHPUnit\AbstractRectorTestCase;
@@ -23,7 +23,7 @@ final class UpdateFileNameByClassNameFileSystemRectorUsingMultipleFilesChangedTr
      */
     public function provideData(): Iterator
     {
-        return $this->yieldFilesFromDirectory(__DIR__ . '/FixtureUsingMultipleFilesChangedTrait');
+        return $this->yieldFilesFromDirectory(__DIR__ . '/Fixture');
     }
 
     public function provideConfigFilePath(): string
