+++ rules-tests/Restoration/Rector/ClassLike/UpdateFileNameByClassNameFileSystemRectorUsingMultipleFilesChangedTrait/config/configured_rule.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+use Rector\Restoration\Rector\ClassLike\UpdateFileNameByClassNameFileSystemRector;
+use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
+
+return static function (ContainerConfigurator $containerConfigurator): void {
+    $services = $containerConfigurator->services();
+    $services->set(UpdateFileNameByClassNameFileSystemRector::class);
+};
